debug

+++ video_inference.py
@@ -116,7 +116,7 @@ class Student():
                 ##### Display new Frame #####
                 prediction[np.where(prediction >= 1)] = 255
                 im = cv2.cvtColor(im, cv2.COLOR_BGR2GRAY)
-                cv2.imshow(self.window_name, np.add(im, prediction)) #prediction.astype(np.uint8)
+                cv2.imshow(self.window_name, np.add(im, prediction).astype(np.uint8)) #prediction.astype(np.uint8)
                 
 
                 ##### Check for New Weights #####
